backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
SQLAlchemy>=2.0
alembic
python-jose[cryptography]
passlib[bcrypt]
email-validator
httpx
pytest
pytest-asyncio
